Use JSChannel for inter-frame communication
We’re doing enough inter-frame communication at this point that it’s
well worth introducing
[JSChannel](http://mozilla.github.io/jschannel/docs/), which provides an
abstraction layer on top of `postMessage`. Helpfully, it includes
two-way communication (send a message and get a response/error back) and
completely takes care of making sure that the message is coming from
where you think it’s coming from.

Allows us to get rid of a ton of boilerplate.