Scoped evaluation of console statement
We want console statements to be evaluated such that each statement is
evaluated in the context of all the previous statements, but variables
defined in the console are _not_ global variables (i.e. they should not
be accessible from the project’s JS code).

Creates a `createScopedEvaluationChain` which does some fancy footwork
to evaluate each statement and return both the value, and a function
that can evaluate the next statement in the context created by the last
one.