Improved scrolling behavior for console
Console now scrolls like most other consoles: once there are enough
entries to overflow the available space, the extra entries overflow the
top of the viewport, with the input stuck to the bottom. You can scroll
up to see older entries.

Implemented with a fairly bananas feature of Flexbox, `flex-direction:
column-reverse`, which just basically turns the contents of the flex
container upside down.