Ensure instructions fit within container
We want the `main` area of the screen to fill up all available vertical
space not needed by the navigation bar, but no more. The instructions
pane should use the vertical space afforded by its parent, but no more;
if it needs more space, it should scroll internally.

Previously we were simply setting the instructions to `overflow: auto`
and crossing our fingers. Chrome interpreted this the way we wanted—keep
the instructions the same height its container naturally wants to be
from flex. However Firefox allowed the instructions to get as big as
they needed to based on content size, pushing the `main` to be bigger
than the way it flexed naturally.

So, add a layer of nesting in the instructions panel; the outer div
takes the height of its container via the default `align-items: stretch`
in the main columns flexbox; then the inner div is absolutely positioned
to fill the outer div and scroll internally.

This is identical to the way that we ensure that the widths of the
editor and output columns are determined solely by flex, without the
possibility of their content affecting the flex calculation.