Gracefully handle compilation failures
In at least one known case, projects currently fail to compile because
of a downstream error in `recast`. In general since it is possible for
project compilation to throw an uncaught error, we want to catch those
errors, display a useful error message, and report the error to Bugsnag.
This does that.