Hide console when errors visible rather than removing it
If we completely remove the console component any time the project is
not in a known valid state, the console will be unmounted and remounted
every time the user types a keystroke in an editor (since this puts it
in a `validating` state for at least a split second). This is
undesirable for several reasons, including that the console input is
focused on mount (meaning it steals focus when the user is typing in a
code editor); and that setting up and tearing down an ACE editor on each
keystroke is a huge waste of runtime resources.

Fixed by just `display: none`ing the console component when it should
not be visible.