Always give imported snapshot new project key
Previously the project key was retained through both project imports and
exports, making it possible for a descendant of a project to overwrite
the original (see fixed issue for details).

Instead, the `snapshotImported` action’s payload contains both a
generated project key and the project data from the snapshot. The
project key, which is generated in the saga, is always used when adding
the project to the store (overwriting the original `projectKey` in the
stored snapshot data).