Explicitly set test NODE_ENV for running tests on Travis
Based on exceptions that Bugsnag is reporting, it seems like the tests
on Travis are being run with a `NODE_ENV` of production, even though the
Karma configuration should be propagating `test` through the Webpack env
plugin. Anyway, it’s probably more sensible to just explicitly set the
environment variable when running tests.

We do keep the env variable set to `production` for the rest of the
Travis run since that’s the correct value for building the production
bundle that’s going to get deployed.