Use a la carte Firebase packages
Instead of pulling in the master Firebase package as a dependency, use a
la carte Firebase packages for the Firebase components that we actually
use—namely `@firebase/app` (the core), `@firebase/auth`, and
`@firebase/database`. Also add `@firebase/app-types` as a dependency to
satisfy the peer dependency of the aforementioned packages.

Just requires changing the imports in the `appFirebase` service to
target the individual packages, and everything seems to work just fine.