Load loop-breaker asynchronously
Previously `loop-breaker` and its `recast` and `esprima` dependencies
were loaded as part of the main bundle, which is entirely unnecessary!
We only need these libraries at all if there is JavaScript to compile
into the project. So, load `loop-breaker` asynchronously, and only if
the project’s JavaScript source is non-empty.