Links from instructions open in a new window
Runs instructions markdown through `remark-external-links`, which by
default adds a `target="_blank"` and `rel="noopener noreferrer"` to
anchor tags.

At this point this is hard-coded into the `toReact` method used here,
with no option to disable it—could always make it optional in the
future, but it’s hard to imagine a situation in which we wouldn’t want
links from processed Markdown to open in a new window.