Upgrade Stylelint to latest
This was a bit of a slog since we heavily patch Stylelint at the webpack
level to get it to behave well in the browser.

Some internal changes to the library broke our old method for doing
this, but after much thrashing the outcome is a happy one, as the
integration now has many fewer patches and hacks. This is thanks mostly
to the discovery of the `checkAgainstRule` function in Stylelint which
is designed for roughly our use case and allows us to bypass loading the
core of the library which tries to configure itself from the filesystem
in lots of different ways.

The only patch needed is to replace the `requireRule` module, which is
used to lazily load rule implementations, with our own version that
greedily loads the one rule we actually use. The original `requireRule`
only works in an environment that allows dynamic `require()` calls—not
the browser.