Regenerate yarn lockfile
The yarn lockfile is path-dependent on the particular order of package
installations and upgrades that we do over time, leading to things like
unnecessarily duplicate packages, sub-dependencies that are out of date,
etc. Apparently the move is to just delete it and re-generate it with
`yarn install`. As long as all direct dependencies are already at their
semver latest, this doesn’t have any effect on the packages we interact
with directly.