Reusable resizable flex HOC
Adds a higher-order component that injects drag-to-resize flex
functionality into any component with a draggable flex layout. All
calculations and state are managed by the HOC, reducing a presentational
component’s responsibilities to forwarding refs, using the flex values
given, and delegating drag events back to the handler prop.

Unlike the previous implementation, the logic here follows a simple rule
that dragging a divider will only affect the size of the regions
adjacent to that divider. While there is no obvious right or wrong
answer here, the new behavior is consistent with other drag-to-resize
implementations I’ve played with, and makes it particularly easy to
support an arbitrary number of flex regions. Anyway, the fundamental
structure of this code could accommodate different logic, so it wouldn’t
be difficult to refine in the future.

The implementation here assumes that the final size of flex regions will
be primarily determined by the `flex-grow` property: the initial main
size represents the _smallest_ that the region should be, and then the
regions grow to fill available space. So, dragging to resize simply
calculates what the ratio of `flex-grow` between the two
divider-adjacent regions should be, and updates them accordingly.

If a drag would cause one of the regions to be smaller than its initial
main size, it is ignored.