Efficient prop updates for resizable flex higher-order component
Ensures we don’t propagate prop updates from the resizable flex HOC
unless something relevant has actually changed.

The main functions in the `resizableFlex` module are independent of the
current state, so we should just define them once and then merge them
into the final props returned by the component.

The props-returning function should return the same instance as
previously if the props themselves are unchanged. To do this, we want to
memoize the function so that new props are only returned if:

* The flex list (received from Redux) has changed
* The `ownProps` (passed directly to the component) have changed

In the latter case, we can’t rely on `ownProps` to be strictly equal,
however; instead, we need to create a custom memoized function that does
a shallow value check (the same kind of check used for pure
`react-redux` containers)