Add linter rule for extraneous imports
`eslint-plugin-import` has a very good rule that fails if we try to
`import` a module that does not exist in `node_modules`. Add this rule
globally, and also add a further rule that forbids dev dependencies from
being imported into modules in the `src` directory.

Makes several changes to the code to follow this rule:

* Fixes a couple of places where we were importing `lodash` instead of
  `lodash-es`
* Adds `lodash.forown` as a dev dependency for use in the gulpfile
* Moves polyfills/shims into the Webpack config (as part of the entry)
  rather than explicitly requiring them in the `src/init` module. This
  is much more consistent with our general approach of treating the code
  in `src` as an idealized environment and pushing all browser and
  language compatibility issues on to the build system.