Use componentDidUpdate instead of componentWillReceiveProps in Editor
The `componentWillReceiveProps` lifecycle hook is deprecated. The
recommendation is to use `componentDidUpdate` instead, which is
straightforward enough, except for components like `<Editor>` which
define `shouldComponentUpdate` as a constant `false` value. This,
however, is not the recommended way to ensure that React never changes a
component’s rendered DOM tree—in fact, `shouldComponentUpdate` may be
treated as a hint in the future.

Instead, prevent any DOM-level updates by making the component’s
`render()` function guaranteed to return the same JSX value every time
(by composing it with lodash’s `constant()` higher-order function).

With this in place, switching to `componentDidUpdate()` is trivial.