Use componentDidMount rather than componentWillMount in Workspace
The `<Workspace>` component reads information out of the query string
and dispatches an action to set up the initial application state based
on what’s in the query (e.g. load a snapshot if the param is there etc.)
This was happening in the `componentWillMount` lifecycle hook, which is
superficially sensible since we would want to fire the Redux action
before the component actually renders.

However, `componentWillMount` is now deprecated, and per the migration
guide:

> There is a common misconception that fetching in `componentWillMount`
> lets you avoid the first empty rendering state. In practice this was
> never true because React has always executed render immediately after
> `componentWillMount`.

So, there is no downside to doing this work in `componentDidMount`.