Share node_modules between Docker and host
There’s not much reason to have node_modules in a volume unless you’re
trying to run Popcode both in a Docker container and on the host (in
which case natively compiled modules could cause a problem). So, just
share `node_modules` between the container and host machine. In
particular this allows us to straightforwardly use `eslint` and friends
in editor integrations.