Upgrade parse5 to 5.0 and use specific package we need
Parse5 5.0.0 breaks the library up into several smaller packages. The
one we need is `parse5-sax-parser`, so swap out the dependency on
`parse5` with one on `parse5-sax-parser`.

https://github.com/inikulin/parse5/blob/master/docs/version-history.md#500

There are substantial breaking changes in the SAXParser API, so update
our usage to fit the new interface.

Finally, these packages now ship in ES6, so run them through the babel
loader.