Notification record
Adds a `Notification` immutable record that encapsulates notifications.
As a corollary, also simplifies the schema of notifications: previously
there was both a `payload` and `metadata`; the `payload` consisted of
data present at the notification’s creation, whereas `metadata` was
potentially updated after the fact. This seems like an unimportant
distinction so I just collapsed everything into `metadata`, which is now
formally a map of key-value pairs which can be added or updated at
notification creation time or later.