Use records for User
Introduces `User` and `UserAccount` records. `User` is always present
and contains information about the current user (logged in or not).
`UserAccount` is present if the user is authenticated and encapsulates
account information.

Login state is now stored in an enum with three potential values:
anonymous, authenticated, or unknown. Unknown is the initial state and
can be used as a signal to not show current login state at all until
Firebase reports it. Actually doing that in a separate commit.