Don’t show login/logout button until initial login state is known
Previously the initial state of the user was to assume they were logged
out unless Firebase reported that they were logged in. This meant that
there was always a flash of **Log in to save** in the top bar which
would then flip to the user account info if they were initially logged
in.

Now we just don’t show anything until we get the initial login state
from Firebase, which is a much nicer experience.