Link GitHub identity to existing account
Adds a **Link GitHub** button to the user menu which is available when
you are logged into an account that does not have an associated GitHub
identity. Clicking this button will initiate a popup-based GitHub
account linking flow.

If the account linking is successful, the application will be in a fully
valid state, allowing the user to use their GitHub credential to e.g.
export gists/repos.

However, while the GitHub credentials _are_ saved to Firebase, they’re
not currently loaded properly for future sessions. This affects both
reloading the page (thus inheriting a preexisting Firebase session) and
logging in to an account that already has a linked GitHub identity.

There is also currently no error handling.