State flow for proposing migration
If a GitHub identity link fails because the identity is linked to
another account, we dispatch an `ACCOUNT_MIGRATION_NEEDED` action, which
initializes a new `AccountMigration` record in the `PROPOSED` state.
This migration record retains the Firebase credential of the account
that would be merged.

The UI should use the existence of a proposed account migration to
display a model confirming that the user wishes to proceed.