Second stage of account migration
Creates a second stage of account migration, the `UNDO_GRACE_PERIOD`,
which is a few seconds during which nothing happens and the user has the
opportunity to stop/cancel the migration before it actually begins.

The `<AccountMigration>` component is now more flexible, using the
migration’s state to determine the header copy, and conditionally
showing a different details sub-component for each stage.