Account migration
Saga initiates account migration, which is fully encapsulated in
Firebase client. Client creates a second SDK instance which is used to
log into the old (GitHub) account using the credential. This instance is
used to retrieve old projects and then save them to the current account.
Then the credential is detached from the old account and attached to the
new one.