Add email and profile scopes to GitHub
Because of what I believe is a bug in the Firebase SDK, we don’t get a
credential back with the auth error unless the response from GitHub
includes an email address. I will file a bug with the SDK, but in the
meantime, requesting `read:user` and `user:email` is a viable workaround.