Handle gapi load failure
If we fail to load `gapi`, for instance if there is no internet
connection, Popcode should still be basically usable. Previously this
situation would throw a promise rejection which was not handled, thus
breaking out of the saga loop and causing the app to stop functioning
properly. Now we handle the rejection and display a message to the user
explaining the situation.