Don’t attempt to support flex resizing in older browsers
Specifically, browsers that don’t implement the modern, un-prefixed
Flexbox standard. It is possible in principle, but would rely on a lot
of browser prefix translation and even translation of different values
and semantics for early flex implementations. At this point it’s not
even clear whether people are using browsers this old; will potentially
revisit this if so.

If the browser is too old to support flex resizing, we avoid showing the
resize mouse cursor, so it does not feel broken; the feature just isn’t there.