Self-contained Docker installation
People should be able to install and run the project by simply running
`docker-compose up`. This makes that work by:

* Making `node_modules` and `bower_components` named volumes
* Running an NPM install as part of the build

We also copy the project root into the container during the build, which
isn’t necessary for development, but would be helpful in a CI
environment.

Finally, we remove the `travis` service and put it in a separate
docker-compose file not under version control.