Use npx as entrypoint for docker container
Prefixing a command with `npx` will effectively add your local
`node_modules/.bin` to the `$PATH` for the purposes of running that
command. Unlike `npm run`, though, it will still run an executable that
lives in another location on your `$PATH`. So, prefixing all commands
run in the Docker container with `npx` should be harmless, and reduces
boilerplate in the Travis build configuration.