Omnibus runtime dependency upgrades
Upgrades an assortment of runtime dependencies:

* `bowser`: `1.4.5` -> `1.9.4`
* `bugsnag-js`: `4.0.3` -> `4.7.3`
* `hast-util-sanitize`: `1.1.1` -> `1.2.0`
* `i18next`: `11.3.2` -> `11.9.0`
* `lodash-es`: `4.17.10` -> `4.17.11`
* `pify`: `3.0.0` -> `4.0.0`
* `remark-external-links`: `2.0.0` -> `3.0.0`
* `shallowequal`: `1.0.2` -> `1.1.0`
* `strip-markdown`: `3.0.0` -> `3.0.2`
* `sweetalert2`: `7.8.6` -> `7.28.2`
* `uuid`: `3.1.0` -> `3.3.2`
* `whatwg-fetch`: `2.0.3` -> `3.0.0`

Per the changelogs, none of the updates involve any changes relevant to
our code, either from a breaking standpoint or a useful new feature
standpoint. Bugsnag does now have a concept of “sessions” which we
explicitly enable in the configuration, although this is also the default.