Batch console.logs every 1 second for infinite loop handling (#1582)
This is an optimization focused on making infinite loop handling more performant for students. The root cause of the performance bottleneck appears to be React's handling of thousands of actions flooding the system during an infinite loop with `console.log` calls inside, which caused the browser UI to lock up before the loop protection code could kick in.

This fixes the bottleneck by adding a debounced queue in front of the log action, so that actions are batched every second and then rendered to the Popcode console.

This replaces the `CONSOLE_LOG_PRODUCED` action with the `CONSOLE_LOG_BATCH_PRODUCED` action.