Explicitly load credentials from Firebase
When Popcode first integrated Firebase, the Firebase client stored
provider credentials and exposed them in the `onAuthStateChange`
callback. However, that changed a long time ago, obliging us to store
provider credentials in the Firebase database.

At the time, I decided to keep the interface of the Firebase client the
same, keeping the storage and retrieval of credentials abstracted from
the caller (in this case, the saga). Ultimately, this was a
path-dependent outcome, and there’s no reason the saga shouldn’t just
handle that task itself.

With an eye toward both becoming more robust to mismatches between
stored credential information and actual linked accounts; and to loading
and storing more information about the user’s provider accounts
(specifically the provider profiles), move handling of credential
persistence into the saga.

While in general we’re able to handle this reasonably elegantly in
the saga, there is one wrinkle, which is that the channel for auth
state changes fires regardless of whether the auth state change is
due to an explicit user action or an implicit change (e.g. the user
logged in/out in a different tab). Oddly, the Firebase client fires the
auth state change handler _before_ resolving the promise returned by
the call to log in.

So, when the user logs in explicitly, we now wait for the promise to
resolve (which gives us a credential we want to store and use), and
in parallel we pull the next auth state change off the channel, simply
discarding it, since it’s redundant with the login promise but contains
less information.