Store linked identity information in objects
Creates a UserIdentityProvider record, which contains information about
a linked user identity (Google, GitHub, etc). Currently this just
stores a single property, which is an `accessToken`. Replaces the
previous approach of storing access tokens as in a map with string
values. The `identityProviders` collection is still a map, with the
keys as the `providerId` (`"github.com"`, `"google.com"`, etc.)

This doesn’t get us anything immediately, but is a prerequisite to two
needed changes:

1. Detect situations where the user has linked an identity provider
   (information that’s available in their Firebase user account object
   when they log in), but we don’t have a stored credential for that
   identity (e.g. a network error prevented us from saving the GitHub
   credential after they linked their account). This can now be
   represented (though is not yet) as a `UserIdentityProvider` with a
   `null` credential.
2. Store additional information about the identities. In particular, we
   want to show the username and avatar for each service in the user
   drop-down, with the option to unlink the identity where appropriate
   (i.e. for GitHub). This would allow students to solve the common
   issue of accidentally linking someone else’s already-logged-in
   GitHub account.