Add buffer to loginState channel
`eventChannel`s are not buffered at all by default; however, we want to
buffer all changes to login state. In particular, the handler for the
first message on the channel (i.e. the initial login state) may itself
change the login state (by signing out in the case that the user logged
in via Firebase no longer has a valid short-lived session cookie). This
caused a bug where the call to sign out from Firebase occurred before
the next call to `take` the login state from the channel, leaving
the user in an indeterminate authentication state.