Load gapi asynchronously before signing out
In the case where we need to instruct `gapi` to open a popup, we have
to do so in the context of a synchronous user action handler. So, we
can’t use the usual approach of awaiting the promise of loading and
configuring `gapi` and then calling the popup on the resolved value.
For that case, we expose `getGapiSync()`, which returns `gapi` if it’s
loaded and throws an error if it isn’t. The idea is that callers to
`getGapiSync()` are responsible for ensuring that `gapi` is definitely
loaded before calling the method.

However, `signOut()` can be called very early in the application
lifecycle, specifically if there is a Firebase session that no longer
has a corresponding short-lived session cookie. Further, we don’t need
to invoke `gapi` synchronously, because signing out does not involve
displaying a popup. This was causing an error where in some cases we
were calling `signOut()`, which was calling methods on a
not-fully-loaded `gapi` instance.

So, fix that issue by switching to the standard
`loadAndConfigureGapi()`, and also change the logic of `loadGapiSync` to
fail fast unless `gapi` is not only loaded but fully configured.