Make most eslint errors warn
We want the build to fail if any linter rules are violated; the most
obvious (and in the past, only) way to do this is to make all rules
errors. However, that is obnoxious when looking at output from
eslint-webpack-plugin and using editor integrations; warnings should be
displayed as warnings. We can now have and eat cake by making most rules
`warn` instead of `error`, and passing the option `--max-warnings=0` to
eslint so that the process still exits with a non-zero code if there is
a warning.