Use promiseRetry to retry gapi load (#1664)
We've been seeing a lot of errors (both user reports and automated error
monitoring) in which the `gapi` script fails to load. Previously we were
using `loadjs`'s built-in retry logic, but this does not do exponential
backoff so will not behave well in a brief network outage. Instead just
use `promiseRetry` like we do everywhere else.