Add validation errors to CodeMirror editor
Uses the `lint` addon, which basically does the right thing. Interface
is a little strange in that you have to give it a function that returns
a collection of errors (you cannot set errors imperatively). So in
effect-world this means updating the editor configuration each time
errors change, providing a new function to return the new set of errors.
This seems to work well enough.