New True Development Environment (#1732)
* New True Development Environment

After facing the fact that Docker adds a substantial performance cost in the
local development environment, build a new One True Development Environment
based on [`nodeenv`](https://github.com/ekalinin/nodeenv) (not to be confused
with `nodenv`, an NVM analog/alternative). Nodeenv is essentially virtualenv
for Node; it allows us to set up an isolated Node environment in a directory
in the project root.

The only dependency of Nodeenv is Python 2.7+, which should already be
installed on nearly any machine someone is using for development.
Installation is quick and painless, and it’s straightforward to reset one’s
environment and rebuild from scratch by just deleting the whole nodeenv
directory.

While `nodeenv` (like virtualenv) has an affordance for “activating” the
environment in the current shell session, thus putting all `nodeenv`
executables on the PATH, we also provide a set of shims in the `tools`
directory to run useful executables like `yarn`, `node`, `npx`, etc. in the
nodeenv without having to activate it in the shell. All `tools/` scripts degrade
gracefully if there is no nodeenv installed, so folks who prefer to maintain
their own Node version can still use the VS Code configurations that invoke
the `tools/` scripts if desired.

* Bump node version to 10.16.1

* Setup script reads Node and Yarn versions from elsewhere

Node version from .node-version; yarn version from package.json

* Experimental PowerShell setup script

* Powershell 5 compatibility

* Use nodeenv PowerShell activation on Windows

* Extract CI steps into scripts; set up script to run them locally

* Make tools/setup.ps1 more rerunnable

* Python setup script

* More organized setup script

* Port all NTDE tools scripts to Python

* Print platform-appropriate path to yarn

* Update task and debug configurations for cross-platform

* Catch the right error

* Remove Bash setup script

* Fixes for Python 2

* Shim scripts fall back to system executables if no nodeenv present

* Exclude nodeenv from jest