Clarify "try" suggestion language
Teaching in a code nation classroom today, a beginner student say the `Try: href="myvalue"` suggestion, and went to paste that _exact_ string into his `img` element. This was an understandable attempt to fix the issue, and I had to correct him that it was showing the _format_ to follow, rather than something specific to add.