Experimental CodeMirror editor (#1670)
* Basic CodeMirror editor component

Creates a basic CodeMirror-backed editor component. Currently hard-coded
to replace the old editor component but should actually be lazy-loaded
only in experimental mode. Lacks most of the features of the old editor
component so far, but bare-bones functionality is there.

* Add validation errors to CodeMirror editor

Uses the `lint` addon, which basically does the right thing. Interface
is a little strange in that you have to give it a function that returns
a collection of errors (you cannot set errors imperatively). So in
effect-world this means updating the editor configuration each time
errors change, providing a new function to return the new set of errors.
This seems to work well enough.

* Add keybinding for autoformatting to CodeMirror editor

* Support text zoom in CodeMirror editor

* CodeMirror editor jumps to error location on click

* Default to not allowing dev dependencies, then whitelist test & build files

* Handle explicit save keyboard shortcut